op-service: move PathFlag to a leaf package so oppprof is importable
op-service/oppprof needs exactly one thing from op-service/flags: the
PathFlag value type, for --pprof.path. But op-service/flags also defines
the --network flag, whose usage string interpolates
chaincfg.AvailableNetworks():

    Usage: fmt.Sprintf("Predefined network selection. Available networks: %s",
        strings.Join(chaincfg.AvailableNetworks(), ", ")),

Go compiles at package granularity, so that one help-string reference put
op-node/chaincfg — and with it op-core/superchain and its embedded
superchain-configs.zip — into the build closure of every consumer of a
pprof flag.

superchain-configs.zip is generated by sync-superchain.sh and gitignored,
so it is absent from the published module. Anything reaching
op-core/superchain therefore cannot be built by a downstream module at all:

    op-core/superchain/chain.go:21:12: pattern superchain-configs.zip:
        no matching files found

That made op-service/oppprof — and so any service wiring up pprof flags —
unimportable outside this repo from v1.19.1 onward, which is when the
registry moved into op-core. Before that the data came from op-geth's own
superchain package, which ships intact, so a downstream op-geth replace
covered it.

PathFlag moves verbatim to the new leaf package op-service/cliflags, which
imports only the standard library. op-service/flags keeps the
chain-selection flags and its chaincfg dependency, which is legitimate.

Adds op-service/deps_test.go, guarding the consumer-facing op-service
packages against a transitive op-core/superchain import via the existing
depguard helper, in the same spirit as op-node/rollup/deps_test.go. The
guard was confirmed to fail on the pre-fix code, reporting the chain
oppprof -> flags -> chaincfg -> op-core/superchain.

Verified: with the bundle deleted, op-service/oppprof compiles, and a
throwaway external module importing op-service/oppprof and
op-service/cliflags builds against this tree. All 15 guarded packages are
bundle-free; `go build ./...` and the touched packages' tests pass.
