+++ op-service/cliflags/pathflag.go
@@ -1,4 +1,15 @@
-package flags
+// Package cliflags holds custom urfave/cli flag value types that are shared
+// across services.
+//
+// It is deliberately a leaf: it must not import anything beyond the standard
+// library and urfave/cli. Custom flag types are reached by low-level packages
+// such as op-service/oppprof, so anything imported here is inherited by every
+// binary and by every downstream consumer that wires up those flags.
+//
+// This is distinct from op-service/flags, which holds OP-Stack chain-selection
+// flags (network, rollup config, fork overrides) and therefore depends on the
+// chain config packages.
+package cliflags
 
 import (
 	"os"
