+++ op-service/deps_test.go
@@ -1,0 +1,63 @@
+package op_service
+
+import (
+	"testing"
+
+	"github.com/ethereum-optimism/optimism/op-service/testutils/depguard"
+)
+
+// consumerFacingPackages are the op-service packages that services — in this
+// repo and in downstream repos — wire up as plumbing: logging, metrics, pprof,
+// RPC, CLI lifecycle. None of them is about chain selection, so none has any
+// business pulling in the superchain config bundle.
+//
+// The list is deliberately not exhaustive. It covers the packages a typical
+// service imports, which is what makes an accidental dependency expensive.
+var consumerFacingPackages = []string{
+	"./cliapp",
+	"./cliflags",
+	"./client",
+	"./ctxinterrupt",
+	"./eth",
+	"./httputil",
+	"./ioutil",
+	"./jsonutil",
+	"./log",
+	"./metrics",
+	"./oppprof",
+	"./rpc",
+	"./signer",
+	"./tls",
+	"./txmgr",
+}
+
+// TestConsumerFacingPackagesAreBundleFree keeps the plumbing half of op-service
+// importable without the superchain config bundle.
+//
+// op-core/superchain embeds superchain-configs.zip, which is generated by
+// sync-superchain.sh and gitignored. Any package in its build closure therefore
+// needs the bundle generated before it will compile — and a downstream module
+// consuming github.com/ethereum-optimism/optimism cannot generate it, because
+// the zip is absent from the published module. Such a package is simply not
+// importable outside this repo.
+//
+// op-service/oppprof used to fail exactly this way: it needs only the PathFlag
+// value type, but took it from op-service/flags, whose --network flag
+// interpolates chaincfg.AvailableNetworks() into its usage string. That one
+// help-string reference pulled op-node/chaincfg, and with it the bundle, into
+// every consumer of a pprof flag. PathFlag now lives in the leaf package
+// op-service/cliflags.
+//
+// op-service/flags is intentionally absent from the list: chain-selection flags
+// legitimately depend on the registry.
+//
+// The check is transitive; a direct-import check would miss the bundle arriving
+// through an intermediate package, which is how it arrived last time.
+func TestConsumerFacingPackagesAreBundleFree(t *testing.T) {
+	for _, pkg := range consumerFacingPackages {
+		t.Run(pkg, func(t *testing.T) {
+			depguard.RequireNoTransitiveImport(t, pkg,
+				"github.com/ethereum-optimism/optimism/op-core/superchain")
+		})
+	}
+}
