+++ op-service/oppprof/cli.go
@@ -7,9 +7,9 @@ import (
 	"strings"
 
 	opservice "github.com/ethereum-optimism/optimism/op-service"
+	"github.com/ethereum-optimism/optimism/op-service/cliflags"
 	"github.com/ethereum-optimism/optimism/op-service/cliiface"
 	openum "github.com/ethereum-optimism/optimism/op-service/enum"
-	"github.com/ethereum-optimism/optimism/op-service/flags"
 	"github.com/urfave/cli/v2"
 )
 
@@ -91,7 +91,7 @@ func CLIFlagsWithCategory(envPrefix string, category string) []cli.Flag {
 		&cli.GenericFlag{
 			Name:     ProfilePathFlagName,
 			Usage:    "pprof file path. If it is a directory, the path is {dir}/{profileType}.prof",
-			Value:    new(flags.PathFlag),
+			Value:    new(cliflags.PathFlag),
 			EnvVars:  opservice.PrefixEnvVar(envPrefix, "PPROF_PATH"),
 			Category: category,
 		},
@@ -131,7 +131,7 @@ func (m CLIConfig) Check() error {
 }
 
 func ReadCLIConfig(ctx cliiface.Context) CLIConfig {
-	profilePathFlag := ctx.Generic(ProfilePathFlagName).(*flags.PathFlag)
+	profilePathFlag := ctx.Generic(ProfilePathFlagName).(*cliflags.PathFlag)
 	return CLIConfig{
 		ListenEnabled:   ctx.Bool(EnabledFlagName),
 		ListenAddr:      ctx.String(ListenAddrFlagName),
